% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Domain HRV Functions .R
\name{Data_Input}
\alias{Data_Input}
\title{Input Data Organization/Cleaning Function}
\usage{
Data_Input(imported_data, input, segment, min, max)
}
\arguments{
\item{imported_data}{Imported csv file.}

\item{input}{RR or MS values.}

\item{segment}{Specifies the upper and lower bounds by minutes.}

\item{min}{If segment is true, what is the minimum value.}

\item{max}{If segment is true, what is the maximum value.}
}
\value{
A dataframe.
}
\description{
This function takes the user's heart rate variation (HRV) data and turns it into a
useable form for the rest of this package. Output from heart rate monitors used in research
are typically either in RR values or MS values. RR values are the milliseconds ellapsed
between successive heart beats, while MS values are a chronological millisecond value that
marks when each heart beat occurs. All statistical values used in HRV take RR values
as an input. Therefore, if the user chooses to upload MS values, the function will
transform it into RR values.
}
\details{
The function allows the user to specify the time frame that they
are interested in, using the "segment" option. The function intakes a minute min/max
value to specify the bounds of their data that they are interested in.
In the literature, the HRV segments less than 30 seconds are not valid and their
use is not advised. Therefor, the program writes a warning if the user chooses to
to use a HRV segment less than 30 seconds.

The function assumes that the user is uploading a csv file with the ms or RR values
in a single row. This matches the type of output given by most heart rate monitoring
software.
}
