% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Domain HRV Functions .R
\name{Descriptive_HRV}
\alias{Descriptive_HRV}
\title{Descriptive Heart Rate -- MIN + MAX + AVERAGE}
\usage{
Descriptive_HRV(data = data)
}
\arguments{
\item{data}{A dataframe with minutes(column 1) and RR values (column 2). Preferably generated by
the Data_Input function.}
}
\value{
List
}
\description{
Gives the user an overview of the data.
HR range (HR max - HR min) is sensitive to Respiratory Sinus Arrhythmia (RSA),
and higher HR range reflects the hearts ability to adapt to the environment.
Lower HR range, in addition to other measures, can be used to diagnose cardiac
risk or as an indicator of distress.
}
