% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Domain HRV Functions .R
\name{NN50_Count}
\alias{NN50_Count}
\title{NN50 Calculation}
\usage{
NN50_Count(data = data)
}
\arguments{
\item{data}{Dataframe with minutes(column 1) and RR values (column 2). Preferably generated by
the Data_Input function.}
}
\value{
Single numeric object
}
\description{
NN50 is the number of inter-beat-intervals that are greater than 50ms.
This is more useful when calculated as a percentage (next function).
}
