% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Domain HRV Functions .R
\name{pNN50_Percentage}
\alias{pNN50_Percentage}
\title{pNN50 Calculation}
\usage{
pNN50_Percentage(data)
}
\arguments{
\item{data}{Dataframe with minutes(column 1) and RR values (column 2). Preferably generated by
the Data_Input function.}
}
\value{
Single numeric object
}
\description{
This value is the percentage of intervals between heart beats that are
greater than 50ms when compared to all heart beats in the segment.
Research suggests that less than 6.25% of heart beats greater than 50ms
indicates an elevated heart rate, putting the individual at cardiac risk or
indicating high levels of distress.
}
