% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Domain HRV Functions .R
\name{RR_Plot}
\alias{RR_Plot}
\title{Plot}
\usage{
RR_Plot(data = data)
}
\arguments{
\item{data}{A dataframe with minutes(column 1) and RR values (column 2). Preferably generated by
the Data_Input function.}
}
\value{
Plot
}
\description{
Gives the user an overview of the data and how heart rate varation changes over time.
}
