% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time Domain HRV Functions .R
\name{SDNN}
\alias{SDNN}
\title{SDNN Function}
\usage{
SDNN(data)
}
\arguments{
\item{data}{Dataframe with minutes(column 1) and RR values (column 2). Preferably generated by
the Data_Input function.}
}
\value{
Single numeric object
}
\description{
SDNN measurement is the standard deviation of all time ellapses heart beats.
This value indicates both parasympathetic and sympathetic nervous system
activity and is imporant measure in cardiology research and practice.
}
